fybot/requirements.txt.txt
telethon
python-telegram-bot==20.0
asyncio
